Projeto funcionando perfeitamente

+++ GalloTube/Interfaces/IVideoTagRepository.cs
@@ -4,15 +4,15 @@ namespace GalloTube.Interfaces;
 
 public interface IVideoTagRepository
 {
-    void Create(int VideoId, byte TagId);
+    void Create(int VideoId, int TagId);
 
-    void Delete(int VideoId, byte TagId);
+    void Delete(int VideoId, int TagId);
 
     void Delete(int VideoId);
 
     List<VideoTag> ReadVideoTag();
 
-    List<Video> ReadVideosByTag(byte TagId);
+    List<Video> ReadVideosByTag(int TagId);
 
     List<Tag> ReadTagsByVideo(int VideoId);
 }
