+++ GalloTube/Models/VideoTag.cs
@@ -11,7 +11,7 @@ public class VideoTag
     public Video Video { get; set; }
 
     [Key, Column(Order = 2)]
-    public byte TagId { get; set; }
+    public int TagId { get; set; }
     [ForeignKey("TagId")]
     public Tag Tag { get; set; }
 }
